// frontend/src/app/generated/model/registerClientRequest.ts
/**
 * Backend API (BFF)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterClientRequest { 
    taxId: string;
    firstName: string;
    lastName: string;
    email: string;
    phone?: string;
    /**
     * Első számla számlaszáma (IBAN)
     */
    accountNumber: string;
    accountType: RegisterClientRequest.AccountTypeEnum;
    currency: RegisterClientRequest.CurrencyEnum;
}
export namespace RegisterClientRequest {
    export const AccountTypeEnum = {
        Checking: 'CHECKING',
        Savings: 'SAVINGS'
    } as const;
    export type AccountTypeEnum = typeof AccountTypeEnum[keyof typeof AccountTypeEnum];
    export const CurrencyEnum = {
        Huf: 'HUF',
        Eur: 'EUR',
        Usd: 'USD'
    } as const;
    export type CurrencyEnum = typeof CurrencyEnum[keyof typeof CurrencyEnum];
}